+++ CHANGELOG.md
@@ -13,6 +13,19 @@ Security - in case of vulnerabilities.
 
 ## [Unreleased] - XXXX-XX-XX
 
+## [1.0.25] - XXXX-XX-XX
+### Fixed
+- Keep approved value on edit organisation
+- Hide public area datasets if organisation is not active or not approved
+- Search and paging on resource type CSV visualization
+- Compare lowercase types before conversion
+
+### Changed
+- Api listDatasets additional array parameter criteria[public] (optional)
+
+### Added
+- Update elastic text keys command
+
 ## [1.0.24] - 2019-03-05
 ### Fixed
 - Improved resource type csv visualization
@@ -409,6 +422,7 @@ Security - in case of vulnerabilities.
 - Search engine fixes
 
 [Unreleased]: https://github.com/governmentbg/data-gov-bg/compare/vUnreleased...HEAD
+[1.0.25]: https://github.com/governmentbg/data-gov-bg/compare/v1.0.24...v1.0.25
 [1.0.24]: https://github.com/governmentbg/data-gov-bg/compare/v1.0.23...v1.0.24
 [1.0.23]: https://github.com/governmentbg/data-gov-bg/compare/v1.0.22...v1.0.23
 [1.0.22]: https://github.com/governmentbg/data-gov-bg/compare/v1.0.21...v1.0.22